Moved trivia setup from bot to trivia
Moved all the trivia question related setup out of bot and into trivia.
Created an Add Await function to Bot, and made use of it in Trivia to
register awaits for answers to questions

renamed the DiscordRB CommandBot in Bot to command_bot (from just bot)
to clear up confusion about which bot is being referenced. Updates to
Messenger were needed for this change